+++ pyproject.toml
@@ -41,7 +41,7 @@ repository = "https://github.com/spacetelescope/stpipe"
 tracker = "https://github.com/spacetelescope/stpipe/issues"
 
 [project.entry-points."asdf.resource_mappings"]
-stpipe = "stpipe.integration:get_resource_mappings"
+stpipe = "stpipe._integration:get_resource_mappings"
 
 [project.scripts]
 stpipe = "stpipe.cli.main:main"
