+++ tests/test_integration.py
@@ -4,7 +4,7 @@ import os
 import asdf
 import yaml
 
-from stpipe.integration import SCHEMAS_PATH
+from stpipe._integration import SCHEMAS_PATH
 
 
 def test_asdf_extension():
